chore: move internal utils to packages/core
