+++ src/nft/components/mint/NFTTotalCost.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode, useCallback, useMemo, useState } from 'react';
 import { infoSvg } from '../../../internal/svg/infoSvg';
-import { multiplyFloats } from '../../../internal/utils/multiplyFloats';
+import { multiplyFloats } from '../../../packages/core/utils/multiplyFloats';
 import { background, border, cn, text } from '../../../styles/theme';
 import { formatAmount } from '../../../token/utils/formatAmount';
 import { useNFTContext } from '../NFTProvider';
