+++ src/packages/core/utils/checkHashLength.test.ts
