+++ src/packages/core/utils/checkHashLength.ts
