+++ src/packages/core/utils/findComponent.test.tsx
