+++ src/packages/core/utils/findComponent.ts
