+++ src/packages/core/utils/formatTokenAmount.test.ts
