+++ src/packages/core/utils/formatTokenAmount.ts
