+++ src/packages/core/utils/multiplyFloats.test.ts
