+++ src/packages/core/utils/multiplyFloats.ts
